fix: separate workflow to four part
