+++ .github/workflows/updateCompanyData_4.yml
@@ -1,0 +1,52 @@
+name: Monthly company data Script Execution
+
+on:
+  schedule:
+    # 每个月2号的00:00 UTC运行脚本
+    - cron: '0 0 2 * *'
+  workflow_dispatch: # 也允许手动触发工作流
+
+jobs:
+  run_python_script:
+    runs-on: ubuntu-latest # 使用 GitHub 提供的 Ubuntu 环境
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3 # 检出仓库代码
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.12' # 设置 Python 版本
+
+      - name: Install Python dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+
+      - name: Start ClickHouse service
+        run: |
+          docker run -d --name clickhouse-server --ulimit nofile=262144:262144 -p 9000:9000 clickhouse/clickhouse-server
+          sleep 30
+
+      - name: List running containers
+        run: docker ps -a
+
+      - name: Check ClickHouse logs
+        run: docker logs clickhouse-server
+
+      - name: Test ClickHouse query
+        run: |
+          docker exec -i clickhouse-server clickhouse-client --query "SELECT 1"
+
+      - name: Run the Python script
+        run: |
+          cd dashboard/company/scripts
+          python workflow_4.py
+        env:
+          DB_HOST: ${{ secrets.DB_HOST }}
+          DB_USER: ${{ secrets.DB_USER }}
+          DB_PASSWORD: ${{ secrets.DB_PASSWORD }}
+          DASHBOARDS_DB_HOST: ${{ secrets.DASHBOARD_DB_HOST }}
+          DASHBOARDS_DB_USER: ${{ secrets.DASHBOARD_DB_USER }}
+          DASHBOARDS_DB_PASSWORD: ${{ secrets.DASHBOARD_DB_PASSWORD }}
